requirements.txt
Django
psycopg2-binary
djangorestframework
dj-database-url>=2.2.0,<3.0
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.2
coverage==7.8.2
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework_simplejwt==5.5.0
idna==3.10
PyJWT==2.9.0
requests==2.32.3
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.4.0